feat: Router 설정 및 폴더 구조화 작업
