+++ src/components/CheckAuth/CheckAuth.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { Outlet } from "react-router-dom";
+
+const CheckAuth = () => {
+  return <Outlet />;
+};
+
+export default CheckAuth;
