+++ src/pages/ClubUpload/ClubUpload.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ClubUpload = () => {
+  return <div>ClubUpload</div>;
+};
+
+export default ClubUpload;
