+++ src/pages/EditProfile/EditProfile.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import ProfileSetting from "../../shared/Profile/ProfileSetting";
+
+const EditProfile = () => {
+  return (
+    <div>
+      <ProfileSetting isBtn={false} />
+    </div>
+  );
+};
+
+export default EditProfile;
