+++ src/pages/PostDetail/PostDetail.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const PostDetail = () => {
+  return <div>PostDetail</div>;
+};
+
+export default PostDetail;
