+++ src/pages/UserFeed/UserFeed.jsx
@@ -1,7 +1,13 @@
-import { HeaderFeed } from "../../shared/Header/HeaderFeed";
-// import Post from "../../share/Post";
-import NoFeed from "../../components/NoFeed/NoFeed";
+import React, { useEffect, useState, useCallback, useRef } from "react";
+import { HeaderBasic } from "../../shared/Header/HeaderBasic";
+import UserProfile from "../../shared/Profile/UserProfile";
+import UserClub from "../../shared/UserClub/UserClub";
+import PostTypeSelectBar from "../../components/PostTypeSelectBar/PostTypeSelectBar";
+import Post from "../../shared/Post/Post";
+import PostAlbum from "../../shared/Post/PostAlbum";
 import Footer from "../../shared/Footer/Footer";
+import ModalInfo from "../../components/ModalModule/ModalInfo";
+import Modal from "../../components/ModalModule/Modal";
 
 const data = {
   post: [
@@ -155,16 +161,88 @@ const data = {
   ],
 };
 
-const Home = () => {
+const UserFeed = () => {
+  const loadingImg = `${process.env.PUBLIC_URL}/assets/img/char-loading-cat.svg`;
+  const [list, setList] = useState(true);
+  const [postDataArray, setPostDataArray] = useState([]);
+
+  const [modal, setModal] = useState(false);
+  const [logout, setLogout] = useState(false);
+  const modalRef = useRef();
+  const handleModalInfo = useCallback(
+    (e) => {
+      setModal(!modal);
+      if (e.target === modalRef.current) {
+        setModal(!modal);
+      }
+    },
+    [modal]
+  );
+  const handleModalLogout = useCallback(() => {
+    setLogout(!logout);
+  }, [logout]);
+  const handleCloseModal = useCallback(() => {
+    setLogout(false);
+  }, []);
+
+  useEffect(() => {
+    if (postDataArray) {
+      const fetchData = async () => {
+        try {
+          await setPostDataArray(data.post);
+        } catch (err) {}
+      };
+      fetchData();
+    }
+  }, [postDataArray]);
+
+  const onListToggle = () => {
+    setList(!list);
+  };
+
   return (
     <div className="page">
-      <HeaderFeed />
+      <HeaderBasic onModalInfo={handleModalInfo} />
       <main>
-        <NoFeed />
+        <UserProfile />
+        <UserClub />
+        <PostTypeSelectBar list={list} onListToggle={onListToggle} />
+        <section className="">
+          <h2 className="ir">유저 게시글</h2>
+          {postDataArray.length === 0 && (
+            <p className="mt-[30%] text-[2.4rem] text-center">아직 생성된 게시글이 없어요 ㅠㅠ</p>
+          )}
+          {postDataArray ? (
+            list ? (
+              postDataArray.map((post, idx) => <Post key={post.id} post={post} />)
+            ) : (
+              <section className="flex flex-wrap my-[1.6rem] mx-[1.6rem]">
+                <h2 className="ir">앨범형</h2>
+                {postDataArray.map((post, idx) => (
+                  <PostAlbum key={post.id} post={post} idx={idx} />
+                ))}
+              </section>
+            )
+          ) : (
+            <img src={loadingImg} alt="잠시만 기다려 주세요." />
+          )}
+        </section>
       </main>
       <Footer />
+      {modal === true ? (
+        <ModalInfo
+          modalRef={modalRef}
+          contentOne="설정 및 개인정보"
+          contentTwo="로그아웃"
+          onModalInfo={handleModalInfo}
+          onModalAction={handleModalLogout}
+        />
+      ) : null}
+      {logout === true ? (
+        <Modal modalRef={modalRef} content="로그아웃하시겠어요?" value="로그아웃" onModalClose={handleCloseModal} />
+      ) : null}
     </div>
   );
 };
 
-export default Home;
+export default UserFeed;
